Add test and refactor a bit tests to simplify a bit
