+++ tests/api.v2/helpers/pipeline/pipelineConstruct.test.js
@@ -8,6 +8,7 @@ const { getQcPipelineStepNames } = require('../../../../src/api.v2/helpers/pipel
 
 const Experiment = require('../../../../src/api.v2/model/Experiment');
 const ExperimentExecution = require('../../../../src/api.v2/model/ExperimentExecution');
+const { createSubsetPipeline } = require('../../../../src/api.v2/helpers/pipeline/pipelineConstruct');
 
 const experimentInstance = new Experiment();
 const experimentExecutionInstance = new ExperimentExecution();
@@ -197,58 +198,6 @@ describe('test for pipeline services', () => {
     expect(createStateMachineSpy.mock.results).toMatchSnapshot();
   });
 
-  it('QC Pipeline is updated instead of created if an error is thrown.', async () => {
-    AWSMock.setSDKInstance(AWS);
-
-    const describeClusterSpy = jest.fn((x) => x);
-    AWSMock.mock('EKS', 'describeCluster', (params, callback) => {
-      describeClusterSpy(params);
-      callback(null, mockCluster);
-    });
-
-    const createStateMachineSpy = jest.fn((stateMachineObject) => _.omit(stateMachineObject, 'definition'));
-    AWSMock.mock('StepFunctions', 'createStateMachine', (params, callback) => {
-      createStateMachineSpy(params);
-      callback({ code: 'StateMachineAlreadyExists' }, null);
-    });
-
-    const updateStateMachineSpy = jest.fn((stateMachineObject) => _.omit(stateMachineObject, 'definition'));
-    AWSMock.mock('StepFunctions', 'updateStateMachine', (params, callback) => {
-      updateStateMachineSpy(params);
-      callback(null, { stateMachineArn: 'test-machine' });
-    });
-
-    const createActivitySpy = jest.fn((x) => x);
-    AWSMock.mock('StepFunctions', 'createActivity', (params, callback) => {
-      createActivitySpy(params);
-      callback(null, { activityArn: 'test-actvitiy' });
-    });
-
-    const startExecutionSpy = jest.fn((x) => x);
-    AWSMock.mock('StepFunctions', 'startExecution', (params, callback) => {
-      startExecutionSpy(params);
-      callback(null, { executionArn: 'test-execution' });
-    });
-
-    experimentInstance.findById.mockReturnValueOnce(
-      { first: () => Promise.resolve(mockExperimentRow) },
-    );
-
-    createQCPipeline.waitForDefinitionToPropagate = () => true;
-
-    await createQCPipeline('testExperimentId', processingConfigUpdate);
-
-    expect(describeClusterSpy).toMatchSnapshot();
-    expect(createStateMachineSpy.mock.results).toMatchSnapshot();
-
-    expect(updateStateMachineSpy).toHaveBeenCalled();
-    expect(updateStateMachineSpy.mock.results).toMatchSnapshot();
-
-    expect(createActivitySpy).toHaveBeenCalled();
-    expect(startExecutionSpy).toHaveBeenCalled();
-    expect(startExecutionSpy.mock.results).toMatchSnapshot();
-  });
-
   it('Create Gem2s pipeline works', async () => {
     AWSMock.setSDKInstance(AWS);
 
@@ -293,8 +242,7 @@ describe('test for pipeline services', () => {
     expect(startExecutionSpy.mock.results).toMatchSnapshot();
   });
 
-
-  it('Gem2s Pipeline is updated instead of created if an error is thrown.', async () => {
+  it('Create Subset pipeline works', async () => {
     AWSMock.setSDKInstance(AWS);
 
     const describeClusterSpy = jest.fn((x) => x);
@@ -303,15 +251,12 @@ describe('test for pipeline services', () => {
       callback(null, mockCluster);
     });
 
-    const createStateMachineSpy = jest.fn((stateMachineObject) => _.omit(stateMachineObject, 'definition'));
+    const createStateMachineSpy = jest.fn(
+      (stateMachineObject) => _.omit(stateMachineObject, ['definition', 'image']),
+    );
+
     AWSMock.mock('StepFunctions', 'createStateMachine', (params, callback) => {
       createStateMachineSpy(params);
-      callback({ code: 'StateMachineAlreadyExists' }, null);
-    });
-
-    const updateStateMachineSpy = jest.fn((stateMachineObject) => _.omit(stateMachineObject, 'definition'));
-    AWSMock.mock('StepFunctions', 'updateStateMachine', (params, callback) => {
-      updateStateMachineSpy(params);
       callback(null, { stateMachineArn: 'test-machine' });
     });
 
@@ -324,22 +269,17 @@ describe('test for pipeline services', () => {
     const startExecutionSpy = jest.fn((x) => x);
     AWSMock.mock('StepFunctions', 'startExecution', (params, callback) => {
       startExecutionSpy(params);
-      callback(null, { executionArn: 'test-execution' });
+      callback(null, { executionArn: 'test-machine' });
     });
 
     experimentInstance.findById.mockReturnValueOnce(
       { first: () => Promise.resolve(mockExperimentRow) },
     );
 
-    createGem2SPipeline.waitForDefinitionToPropagate = () => true;
-
-    await createGem2SPipeline('testExperimentId', taskParams);
-
+    await createSubsetPipeline('testExperimentId', taskParams);
     expect(describeClusterSpy).toMatchSnapshot();
-    expect(createStateMachineSpy.mock.results).toMatchSnapshot();
 
-    expect(updateStateMachineSpy).toHaveBeenCalled();
-    expect(updateStateMachineSpy.mock.results).toMatchSnapshot();
+    expect(createStateMachineSpy.mock.results).toMatchSnapshot();
 
     expect(createActivitySpy).toHaveBeenCalled();
     expect(startExecutionSpy).toHaveBeenCalled();
